// solauto-sdk/src/generated/types/positionTokenUsage.ts
/**
 * This code was AUTOGENERATED using the kinobi library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun kinobi to update it.
 *
 * @see https://github.com/metaplex-foundation/kinobi
 */

import { PublicKey } from '@metaplex-foundation/umi';
import {
  Serializer,
  array,
  bytes,
  publicKey as publicKeySerializer,
  struct,
  u16,
  u64,
  u8,
} from '@metaplex-foundation/umi/serializers';
import { TokenAmount, TokenAmountArgs, getTokenAmountSerializer } from '.';

export type PositionTokenUsage = {
  mint: PublicKey;
  decimals: number;
  padding1: Array<number>;
  amountUsed: TokenAmount;
  amountCanBeUsed: TokenAmount;
  baseAmountMarketPriceUsd: bigint;
  flashLoanFeeBps: number;
  borrowFeeBps: number;
  padding2: Array<number>;
  padding: Uint8Array;
};

export type PositionTokenUsageArgs = {
  mint: PublicKey;
  decimals: number;
  padding1: Array<number>;
  amountUsed: TokenAmountArgs;
  amountCanBeUsed: TokenAmountArgs;
  baseAmountMarketPriceUsd: number | bigint;
  flashLoanFeeBps: number;
  borrowFeeBps: number;
  padding2: Array<number>;
  padding: Uint8Array;
};

export function getPositionTokenUsageSerializer(): Serializer<
  PositionTokenUsageArgs,
  PositionTokenUsage
> {
  return struct<PositionTokenUsage>(
    [
      ['mint', publicKeySerializer()],
      ['decimals', u8()],
      ['padding1', array(u8(), { size: 7 })],
      ['amountUsed', getTokenAmountSerializer()],
      ['amountCanBeUsed', getTokenAmountSerializer()],
      ['baseAmountMarketPriceUsd', u64()],
      ['flashLoanFeeBps', u16()],
      ['borrowFeeBps', u16()],
      ['padding2', array(u8(), { size: 4 })],
      ['padding', bytes({ size: 32 })],
    ],
    { description: 'PositionTokenUsage' }
  ) as Serializer<PositionTokenUsageArgs, PositionTokenUsage>;
}